% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{add_mean_and_sd}
\alias{add_mean_and_sd}
\title{Adds the mean values and multiplies the data by the standard deviations
previously estimated from the training set.}
\usage{
add_mean_and_sd(x, means, sds)
}
\arguments{
\item{x}{Matrix, dataset to be "unscaled".}

\item{means}{Numeric vector, mean values to be added.}

\item{sds}{Numeric vector, standard deviations to be rescaled.}
}
\description{
Adds the mean values and multiplies the data by the standard deviations
previously estimated from the training set.
}
