% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{apply_center_and_scale}
\alias{apply_center_and_scale}
\title{Centers and scales the test data using the values estimated
from the training set.}
\usage{
apply_center_and_scale(bara_fit, x)
}
\arguments{
\item{bara_fit}{Object of class BaraFit, created using the training set.}

\item{x}{Matrix, data to be scaled.}
}
\value{
Matrix, the input data  scaled using the values in bara_fit.
}
\description{
Centers and scales the test data using the values estimated
from the training set.
}
