% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{bara_adjust}
\alias{bara_adjust}
\title{Adjusts a test set using the set of reference samples}
\usage{
bara_adjust(bara_fit, x, ref, batch = NULL, verbose = TRUE,
  force = FALSE)
}
\arguments{
\item{bara_fit}{Object of class BaraFit, created using the training set.}

\item{x}{Matrix, the test data set with samples in rows.}

\item{ref}{Numeric vector, index of the reference samples in the test set.}

\item{batch}{Character/Numeric/Factor. The vector should represent
the batch belongings for the samples in the test set x. If a vector is
supplied, each batch is independently adjusted. Therefore, each batch
must contain at least 1 reference sample.}

\item{verbose}{Logical, should runtime messages be printed? Default is TRUE.}

\item{force}{Logical, should errors associated to the identification of 
batches without reference samples be suppressed? Default is \code{FALSE}.}
}
\value{
A matrix corresponding to the normalized test set.
}
\description{
Adjusts a test set using the set of reference samples
}
