% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{bara_fit}
\alias{bara_fit}
\title{Creates a BARA fit object}
\usage{
bara_fit(x, ref, ndim = NULL, loss = NULL, scale_var = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{Matrix, the training data set with samples in rows.}

\item{ref}{Numeric vector, index of the reference samples.}

\item{ndim}{Integer, describes the number of SVD dimensions to retain.
If NULL, the all possible dimensions are calculated.}

\item{loss}{Numeric between 0 and 1. Describes the maximum amount of
variance that can be lost during compression.}

\item{scale_var}{Logical, should the variance be scaled? Default is FALSE.}

\item{verbose}{Logical, should runtime messages be printed? Default is TRUE.}
}
\value{
An object of class BaraFit.
\describe{
  \item{\env{x}}{Compressed and reconstructed training data set.
    Can be used to define subsequent prediction model.}
  \item{\env{ref}}{Index of reference samples.}
  \item{\env{means}}{Mean values of the columns of the input matrix x.}
  \item{\env{sds}}{Standard deviations of the columns of the input matrix x.}
  \item{\env{right_vectors}}{Right singular vectors of the input matrix x.}
  \item{\env{ref_means}}{Mean values of the reference samples in the
      compressed data space.}
  \item{\env{call}}{Function call.}
}
}
\description{
The function is applied to the training set and returns a BaraFit
object that can be used to adjust batch effects in subsequently acquired
test sets.
}
