% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_dimensions.R
\name{bara_optimize_ext}
\alias{bara_optimize_ext}
\title{This function can be used to optimize the number of dimensions to
retain when normalizing with BARA.}
\usage{
bara_optimize_ext(x_train, ref_train, y_train, x_test, ref_test, y_test,
  fit_fun, pred_fun, perf_fun = NULL, batch_test = NULL,
  perf_objective = c("maximize", "minimize"), max_dim = 100,
  frac_rm = 0.05, fit_args = NULL, pred_args = NULL,
  perf_args = NULL, scale_var = FALSE, verbose = TRUE, cores = 1,
  .export = NULL, .packages = NULL, seed = 16438)
}
\arguments{
\item{x_train}{Matrix, training data set. Samples in rows.}

\item{ref_train}{Numeric vector, index of reference samples in the
training set.}

\item{y_train}{Response variables corresponding to the training set.}

\item{x_test}{Matrix, test data set. Samples in rows.}

\item{ref_test}{Numeric vector, index of reference samples in the
test set.}

\item{y_test}{Response variable corresponding to the test set.}

\item{fit_fun}{Function, fits a prediction object from the training
data. The function must take the following arguments:
\itemize{
\item \code{x}:  A matrix corresponding to the compressed training data.
\item \code{y}:  The response variable corresponding to y_train.
\item \code{...}:  See details for additional arguments
  passed to the function.
}}

\item{pred_fun}{Function, classifies the samples in the test set and
returns the predictions. The function must have the following
parameters:
\itemize{
  \item \code{object}:  The object returned by \code{fit_fun}.
  \item \code{x}:  A matrix corresponding to the compressed test set.
  \item \code{...}:  See details for additional arguments.
}}

\item{perf_fun}{Function, estimates the performance of the generated
    predictions. The performance should be returned as a numeric value.
    This parameter can also be left to the default value of NULL if no
    performance scores should be calculated. If a function is passed.
    it must have the the following parameters:
    \itemize{
      \item \code{y}: The response variable corresponding to the
        \code{y_test} argument.
      \item \code{pred}: The predictions generated with \code{pred_fun}.
      \item \code{...}:  See details for additional arguments.

    }}

\item{batch_test}{Character/numeric/factor, describing the batch
belongings of the test set.}

\item{perf_objective}{Character, either \code{"maximize"} or
\code{"minimize"}. Should the performance metric be maximized
or minimized?}

\item{max_dim}{Numeric, determines the maximum number of dimensions
to retain. This parameter can be decreased to reduce computational
time. Default value is \code{100}.}

\item{frac_rm}{Numeric between 0 and 1. How large fraction of the dimensions
should be dropped in each iteration. For example, if
 \code{frac_rm = 0.05}, the number of dimensions retained decreases with
 5 percent in each iteration.}

\item{fit_args}{List with arguments passed to \code{fit_fun}.}

\item{pred_args}{List with arguments passed to \code{pred_fun}.}

\item{perf_args}{List with arguments passed to \code{perf_fun}.}

\item{scale_var}{Logical, should the variances be standardized.
The default value is \code{FALSE}}

\item{verbose}{Logical, should runtime messages be displayed? Default
value is \code{TRUE}}

\item{cores}{Integer, number of cores to use for computations.}

\item{.export}{Variables that any of \code{fit_fun}, \code{pred_fun}
or \code{fit_fun} depends on. Any input is passed on to foreach.}

\item{.packages}{Packages that any of \code{fit_fun}, \code{pred_fun}
or \code{fit_fun} depends on. Any input is passed on to foreach.}

\item{seed}{Integer, seed for reproducible computations.}
}
\value{
An object of class BaraOpt containing:
    \itemize{
      \item \code{fit}: A BaraFit object with the optimal dimensions retained.
        \code{best_dim} Numeric, the number of dimensions retained for the
            optimal prediction performance.
      \item \code{iterations}: List, results from the iterative process,
          including predictions and prediciton models for the number of
          dimensions examined.

    }
}
\description{
This function is used to estimate the optimal number of dimensions
to retain when normalizing with BARA. The optimization is performed using
an external test set, ideally affected by batch effects.
}
