% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{center_and_scale}
\alias{center_and_scale}
\title{Centers and scales the matrix and returns mean and sd vectors.}
\usage{
center_and_scale(x, scale_var = FALSE)
}
\arguments{
\item{x}{Matrix, the training data set with samples in rows.}

\item{scale_var}{Logical, should the variance be scaled? Default is FALSE.}
}
\description{
Centers and scales the matrix and returns mean and sd vectors.
}
