% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{check_matrix}
\alias{check_matrix}
\title{Makes sure that x is a numeric matrix.}
\usage{
check_matrix(x, verbose = TRUE)
}
\arguments{
\item{x}{Input data, should be a matrix.}

\item{verbose}{Logical, should messages be printed? Default is TRUE.}
}
\description{
Makes sure that x is a numeric matrix.
}
