% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{check_ref}
\alias{check_ref}
\title{Validates the format of the reference samples}
\usage{
check_ref(ref, n_samples, verbose = TRUE)
}
\arguments{
\item{ref}{Numeric, reference samples.}

\item{n_samples}{Numeric, number of samples in the data matrix.}

\item{verbose}{Logical, should runtime messages be printed? Default is TRUE.}
}
\description{
Validates the format of the reference samples
}
