% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{check_svd_params}
\alias{check_svd_params}
\title{Get SVD dimension identification type}
\usage{
check_svd_params(x, ndim, loss)
}
\arguments{
\item{x}{Matrix, samples in rows.}

\item{ndim}{Integer, describes the number of SVD dimensions to retain.
If NULL, the number of dimensions is based on loss. If both ndim
and loss is NULL, all dimensions are returned.}

\item{loss}{Numeric between 0 and 1. Describes the maximum amount of
variance that can be lost during compression.}
}
\description{
Get SVD dimension identification type
}
