% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{compress_data}
\alias{compress_data}
\title{Compress a dataset}
\usage{
compress_data(right_vectors, x)
}
\arguments{
\item{right_vectors}{Matrix, the right singular vectors of the singular
value decomposition performed on the training set.}

\item{x}{Matrix, a dataset to be projected onto singular vectors.}
}
\value{
Matrix of the projected dataset.
}
\description{
Compress a dataset
}
