% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{generate_lgl}
\alias{generate_lgl}
\title{Generate a logical vector from numeric vector}
\usage{
generate_lgl(true, n_samples)
}
\arguments{
\item{true}{Numeric vector, index of true indices.}

\item{n_samples}{Numeric, number of samples in the data (rows of matrix x).}
}
\value{
Logical vector where TRUE entries represent reference samples.
}
\description{
Generate a logical vector from numeric vector
}
