% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_dimensions.R
\name{get_best_dimension}
\alias{get_best_dimension}
\title{Determines which dimensions optimizes the performance score.}
\usage{
get_best_dimension(perf, dims, objective)
}
\arguments{
\item{perf}{Numeric vector, represents the performance scores.}

\item{dims}{Numeric vector, describes the number of dimensions retained.}

\item{objective}{Character, one of \code{maximize} or \code{minimize}.
Determines if the performance score should be maximized or minimized.}
}
\value{
The lowest number of dimensions retained to achive the optimal
    performance score.
}
\description{
Determines which dimensions optimizes the performance score.
}
