% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_dimensions.R
\name{get_dim_sizes}
\alias{get_dim_sizes}
\title{Calculates the number of dimensions to retain given a
fraction decrease at each iteration}
\usage{
get_dim_sizes(frac_rm, n_var)
}
\arguments{
\item{frac_rm}{Numeric between 0 and 1. Determines how many
dimensions to dropped in each iteration.}

\item{n_var}{Numeric, number of variables in the dataset.}
}
\value{
Numeric vector describing the number of dimensions to
    retain in each iteration.
}
\description{
Calculates the number of dimensions to retain given a
fraction decrease at each iteration
}
