% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{reconstruct_data}
\alias{reconstruct_data}
\title{Reconstructs the data to its original dimension}
\usage{
reconstruct_data(right_vectors, pcs)
}
\arguments{
\item{right_vectors}{Matrix, the right singular vectors of the singular
value decomposition performed on the training set.}

\item{pcs}{Matrix, a compressed dataset to be reconstructed.}
}
\value{
Matrix of the reconstructed dataset.
}
\description{
Reconstructs the data to its original dimension
}
