% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{squish_train}
\alias{squish_train}
\title{Centers, possibly scales, compresses, reconstructs and restores
scale and mean of the training set}
\usage{
squish_train(bara_fit, x)
}
\arguments{
\item{bara_fit}{Object of class BaraFit, created using the training set.}

\item{x}{Matrix, data to be processed without adjustment}
}
\value{
Matrix of compressed and reconstructed input x.
}
\description{
Centers, possibly scales, compresses, reconstructs and restores
scale and mean of the training set
}
