% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{truncate_bara_fit}
\alias{truncate_bara_fit}
\title{Decreases the number of dimensions retained in a BaraFit object}
\usage{
truncate_bara_fit(bara_fit, x, dim_to_keep)
}
\arguments{
\item{bara_fit}{An object of class BaraFit.}

\item{x}{Matrix, the training set used to define the original BaraFit
object.}

\item{dim_to_keep}{Numeric, describes the number of dimensions to retain.}
}
\value{
An object of class BaraFit with a decreased number of dimensions
    retained.
}
\description{
Decreases the number of dimensions retained in a BaraFit object
}
