% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bara.R
\name{check_batch}
\alias{check_batch}
\title{Verifies the input format of the batch arguments.}
\usage{
check_batch(batch, ref, n_samples)
}
\arguments{
\item{batch}{Character/Numeric/Factor. The vector should represent
the batch belongings for the samples in the test set x. If a vector is
supplied, each batch is independently adjusted. Therefore, each batch
must contain at least 1 reference sample.}

\item{ref}{Numeric vector, index of the reference samples in the test set.}

\item{n_samples}{Numeric, number of samples in x.}
}
\value{
Character vector.
}
\description{
Verifies the input format of the batch arguments.
}
